refactor(analysis-types): extract complexity histogram DTO
