+++ ci/proof.toml
@@ -346,7 +346,10 @@ coverage = true
 [[scope]]
 name = "analysis_types_complexity"
 kind = "rust"
-paths = ["crates/tokmd-analysis-types/src/complexity.rs"]
+paths = [
+  "crates/tokmd-analysis-types/src/complexity.rs",
+  "crates/tokmd-analysis-types/src/complexity/**",
+]
 packages = ["tokmd-analysis-types"]
 proof = ["cargo test -p tokmd-analysis-types complexity --verbose"]
 mutation = true
